backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pydantic[email]==2.12.5
python-multipart==0.0.26
razorpay==2.0.1
python-dotenv==1.2.2
bcrypt==5.0.0
